Version Beta Final

+++ tienda/forms.py
@@ -2,9 +2,10 @@ from django import forms
 
 class FormularioProducto(forms.Form):
     nombre = forms.CharField()
-    precio = forms.IntegerField()
+    precio = forms.DecimalField()
     descripcion = forms.CharField()
     categoria = forms.CharField()
+    
 
 class FormularioBusqueda(forms.Form):
     termino = forms.CharField(max_length=100)
